agent: split the two oversized service files (pure movement)
Phase 5 of design/service-reliability.md, deliberately last so phases 1-4
landed as small diffs on the layout reviewers already knew. No behavior
change, no API change, no output change.

internal/cli/agent.go (2069 lines) is removed; its contents become six
files named for what they are:

  servicerun.go       the daemon mode `jit service run` — process
                      lifetime, log discipline, TTL validation
  servicelaunchd.go   every launchd interaction: plist write/read,
                      reload+kickstart, install, the launchctl print parse
  servicecmds.go      the subcommands a human types, plus unlock/lock
                      and the init() wiring
  servicestatus.go    `jit service status` and its renderers
  servicelog.go       `jit service log` (rendering stays in agentlogview.go)
  agentclient.go      how every other command reaches the service, and
                      the single source of what to say when it can't

internal/agent/server.go (2062 -> 687) sheds four:

  fetcher.go     the MEK-fetcher seam — where key material enters
  transport.go   the socket boundary: listen, accept, decode, bound
  session.go     the whole session state machine, so the mu /
                 challengeMu discipline is readable in one place
                 instead of reconstructed across 2000 lines
  events.go      the event trail and the use-aggregation windows

humanAgo moves to format.go rather than a service file: `migrate undo`
shares it. agent_status_test.go's launchd-seam tests move to
servicelaunchd_test.go and its two command-level tests to
servicecmds_test.go, next to what they exercise.

Verified as pure movement rather than asserted: the set of top-level
declarations before and after is identical per package, checked
mechanically, and the full -race suite, vet, gofmt, staticcheck, gosec
and docs-gen all pass unchanged. Exactly one test needed editing, and
only because it names source files: the agentPlistNeedsRepoint source
guard now reads servicecmds.go instead of agent.go.
