+++ design/service-reliability.md
@@ -1,13 +1,13 @@
 # Service reliability: demand-spawned lifecycle, honest health, a broker that keeps its promises
 
-**Status: phases 1-4 implemented, 2026-08-17 (phases 1-2 on branch
-service-reliability, PR #72; phase 3 stacked as service-log-hygiene,
-PR #73; phase 4 stacked as service-broker-hardening, PR #74; both
-wording previews approved by eye the same day). Phase 5 (the
-mechanical file split) is specified but not yet built. Drafted from a
-full review of the service surface (three parallel code reviews over
-`internal/agent`, `internal/cli/agent*.go` and the log/history/status
-surfaces, 43 findings) plus a verified production incident.**
+**Status: fully implemented, 2026-08-17, as a stack of four PRs — #72
+(phases 1-2, branch service-reliability), #73 (phase 3,
+service-log-hygiene), #74 (phase 4, service-broker-hardening), #75
+(phase 5, service-file-split). Both wording previews were approved by
+eye the same day. Drafted from a full review of the service surface
+(three parallel code reviews over `internal/agent`,
+`internal/cli/agent*.go` and the log/history/status surfaces, 43
+findings) plus a verified production incident.**
 
 ## The incident that forced this
 
@@ -301,6 +301,22 @@ recordUse family), with `server.go` keeping the struct, hooks and
 dispatch. The struct's three locking regimes are currently
 reconstructed by scanning 1760 lines; the split makes each regime's
 home explicit.
+
+**As implemented.** `agent.go` was removed outright (its six
+successors cover it); `humanAgo` went to `format.go` rather than a
+service file, since `migrate undo` shares it. Result: `agent.go` 2069
+→ six files of 173-546, and `server.go` 2062 → 687 plus four files of
+70-853. `session.go` is the largest at 853 rather than the estimated
+450 — the phase-4 disclosed-prompt backoff landed there, and it is
+still one concern (the session state machine) with one locking
+regime, so it was left whole rather than split on size alone.
+
+Verified as pure movement rather than asserted: the set of top-level
+declarations before and after is byte-identical per package, checked
+mechanically, and every gate passes unchanged. One test needed
+editing, and only because it names source files — the
+`agentPlistNeedsRepoint` source guard now reads `servicecmds.go`
+instead of `agent.go`.
 
 ## Test plan
